removed initiativeHandler.js, added id to localstorage
