+++ index.html
@@ -56,6 +56,5 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
     <script src="./assets/js/homepage.js" type="module"></script>
     <script src="./assets/js/modules/battleModal.js" type="module"></script>
-    <script src="./assets/js/modules/initiativeHandler.js" type="module"></script>
 </body>
 </html>
